pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "bookingassistant"
version = "0.1.0"
dependencies = [
    "aiogram",
    "python-dotenv",
    "dateparser",
    "aiohttp",
    "aioresponses",
    "pytest-asyncio",
    "asyncpg",
    "fpdf2",
    "certifi",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
